requirements.txt
langchain
google-generativeai
flask
langchain-google-genai